Mark district 2 with a text label at the default zoom level of 13
When a user lands on the page, we want them to know the distinction between the different heatmap colours on the map, so we start them off with a label mentioning 'District 2'. When they zoom in, this label disappears, allowing them to interact with the map, and when they zoom out to level 13, the label re-appears, allowing them to understand the significance of the different colours on the heatmap.